pyproject.toml
[project]
name = "ai-rota-system"
version = "0.1.0"
description = "AI-based rota system for healthcare"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "openai>=1.3.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.4.0",
    "python-dotenv>=1.0.0",
    "xlrd>=2.0.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0"
] 